+++ src/tests/hooks/useViewportIntersection.spec.tsx
@@ -3,71 +3,7 @@ import React from 'react';
 
 import { ViewportIntersectionProvider, useViewportIntersection } from 'src/hooks/useViewportIntersection';
 
-class MockIO {
-  elements: Set<Element> = new Set();
-  static visibleElements: Set<Element> = new Set();
-  static activeIOs: Set<MockIO> = new Set();
-
-  constructor(
-    public callback: IntersectionObserverCallback,
-    public options?: IntersectionObserverInit,
-  ) {
-    MockIO.activeIOs.add(this);
-  }
-
-  observe(e: Element) {
-    if (!(e instanceof Element)) {
-      throw new TypeError('not an element');
-    }
-    this.elements.add(e);
-    this.changeVisibility(e, MockIO.visibleElements.has(e));
-  }
-
-  unobserve(e: Element) {
-    if (!(e instanceof Element)) {
-      throw new TypeError('not an element');
-    }
-    this.elements.delete(e);
-  }
-
-  private changeVisibility(e: Element, visible: boolean) {
-    const entry = {
-      target: e,
-      intersectionRatio: visible ? 1 : 0,
-      isIntersecting: visible,
-    };
-    this.callback([entry as any], this as unknown as IntersectionObserver);
-  }
-
-  static reveal(e: Element) {
-    if (MockIO.visibleElements.has(e)) {
-      return;
-    }
-    MockIO.visibleElements.add(e);
-    for (const io of MockIO.activeIOs.values()) {
-      if (io.elements.has(e)) {
-        io.changeVisibility(e, true);
-      }
-    }
-  }
-
-  static hide(e: Element) {
-    if (!MockIO.visibleElements.has(e)) {
-      return;
-    }
-    MockIO.visibleElements.delete(e);
-    for (const io of MockIO.activeIOs.values()) {
-      if (io.elements.has(e)) {
-        io.changeVisibility(e, false);
-      }
-    }
-  }
-
-  disconnect() {
-    this.elements.clear();
-    MockIO.activeIOs.delete(this);
-  }
-}
+import MockIO, { prepareTest } from '../utils/MockIO';
 
 describe('useViewportIntersection', () => {
   function TestComponent(props: { callback: (visible: boolean) => void }) {
@@ -77,17 +13,7 @@ describe('useViewportIntersection', () => {
     );
   }
 
-  let originalIO: typeof MockIO;
-
-  beforeAll(() => {
-    originalIO = window.IntersectionObserver as unknown as typeof MockIO;
-    window.IntersectionObserver = MockIO as unknown as typeof IntersectionObserver;
-  });
-
-  afterAll(() => {
-    window.IntersectionObserver = originalIO as unknown as typeof IntersectionObserver;
-  });
-
+  prepareTest();
   afterEach(() => {
     act(() => {
       cleanup();
